+++ tests/WeightTest.php
@@ -14,4 +14,14 @@ class WeightTest extends TestCase
 
         $this->assertEquals(220.4623, $lbs);
     }
+
+    /**
+     * @test
+     */
+    public function it_can_convert_kilograms_to_grams()
+    {
+        $grams = Weight::fromKilograms(1)->toGrams();
+
+        $this->assertEquals(1000, $grams);
+    }
 }