+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `unit-conversions` will be documented in this file
 
+## 1.1.0 - 2020-09-24
+
+- Length Conversions:
+  - Convert meters to kilometers feature.
+  - Convert meters to miles feature.
+
 ## 1.0.0 - 2020-09-24
 
 - Initial release